pyproject.toml
[project]
name = "raspi-recipe-app"
version = "0.1.0"
description = "Lightweight recipe finder/saver with LLM and FastAPI for Raspberry Pi 3"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.112.0,<1.0.0",
    "uvicorn[standard]>=0.30.0,<1.0.0",
    "jinja2>=3.1.4,<4.0.0",
    "httpx>=0.27.0,<1.0.0",
    "pydantic>=2.7.0,<3.0.0",
    "python-dotenv>=1.0.1,<2.0.0",
    "python-multipart>=0.0.9,<1.0.0",
    "google-genai>=1.30.0",
]

[tool.uv]
package = true
